+++ src/components/custom/signup.js
@@ -6,8 +6,7 @@ import { FlexGridContainer } from 'components/containers'
 import { Header, Text, Image, StyledLink } from 'components/elements'
 import { localize, Localize } from 'components/localization'
 import Wrapper from 'components/containers/wrapper'
-import Button from 'components/form/button'
-import Input from 'components/form/input'
+import { Button, Input } from 'components/form'
 import validation from 'common/validation'
 import TrafficSource from 'common/traffic-source'
 import { LocalStore } from 'common/storage'
