+++ src/components/form/index.js
@@ -1,0 +1,4 @@
+import Button from './button'
+import Input from './input'
+
+export { Button, Input }
