+++ src/pages/dmt5/_hero.js
@@ -5,7 +5,7 @@ import { Container, Show } from 'components/containers'
 import device from 'themes/device.js'
 import { Header, Modal, useModal, Image } from 'components/elements'
 import SignupModal from 'components/custom/signup-modal'
-import Button from 'components/form/button.js'
+import { Button } from 'components/form'
 import dmt5_logo from 'images/svg/dmt5-32.svg'
 
 const HeroContent = styled.div`
