+++ src/pages/dtrader/_experience.js
@@ -6,7 +6,7 @@ import device from 'themes/device.js'
 import { Modal, useModal, Header, Text } from 'components/elements'
 import SignupModal from 'components/custom/signup-modal'
 import { deriv_app_url } from 'common/utility'
-import Button from 'components/form/button'
+import { Button } from 'components/form'
 
 const StyledSection = styled(SectionContainer)`
     background-color: var(--color-grey-1);
