+++ src/pages/home/_hero.js
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { HeroHeader, StyledHeader } from './_headers'
 import device from 'themes/device'
-import Button from 'components/form/button'
+import { Button } from 'components/form'
 import { Container, Show } from 'components/containers'
 import { Header, Modal, useModal } from 'components/elements'
 import { localize } from 'components/localization'
