+++ src/pages/reset-password/index.js
@@ -5,8 +5,7 @@ import Layout from 'components/layout/layout'
 import { localize, WithIntl } from 'components/localization'
 import { Container, SEO } from 'components/containers'
 import { Header, Text } from 'components/elements'
-import Input from 'components/form/input'
-import Button from 'components/form/button'
+import { Input, Button } from 'components/form'
 import validation from 'common/validation'
 import { BinarySocketBase } from 'common/websocket/socket_base'
 import Login from 'common/login'
